mkosi: drop systemd-repart full config and specify fs type
We just want to change the default systemd-repart fs form erofs to
squashfs: for this reason there is no need to re-declare the full
systemd-repart configuration, we can just pass the
SYSTEMD_REPART_OVERRIDE_FSTYPE_ROOT=squashfs
environment variable in the Build section.

We still cannot move to the 'sysext' Output format (where the
SYSTEMD_REPART_OVERRIDE_FSTYPE_ROOT is honored) as we want to fully
control the content of the produced extension-image.NAME file.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>